requirements.txt
requests>=2.31,<3
pandas>=2.1,<3
schedule>=1.2,<2
python-dotenv>=1.0,<2
flask>=3.0,<4
google-cloud-storage>=2.10,<3
